% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{formatDate}
\alias{formatDate}
\title{formatDate}
\usage{
formatDate(x, format = "1997-04-23 14:53:42")
}
\arguments{
\item{x, characte}{or date or POSIXt}
}
\description{
format date using Go stytle
}
\examples{
formatDate('2016-12-23 11:23:45','1997-04-23')
}
\author{
HatMatrix
}

